### 2023-05-03 #5366 @jansim
Thank you for suggesting some reviewers @oliviaguest !

This is our first time submitting to JOSS, so if there's a way I can help with the process, I'd be happy to 😊
