### 2023-05-21 #5366 @crvernon
:wave: Alright @jansim let's get you some reviewers!  Would you mind making a few recommendations for potential reviewers for this work?  You can also check out [this list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Just post some suggestions in this thread without the `@` and I'll get on the hustle.

Thanks for your patience and have a great day!
