### 2023-05-24 #5366 @jansim
Hi @crvernon ,

Thank you for agreeing to be an editor! 🙏

Regarding potential reviewers I think `danielruss` would be an excellent fit, as he developed a similar solution.

From the list of JOSS reviewers I think the following could make sense, everyone here has R as their preferred language and a rough match in terms of their areas of expertise (typically ML / stats).

- `vpergim`
- `ben-aaron188`
- `welch16`
- `bernhardclemm`
- `n8thangreen`

I also received many suggestions from within our research group, but was unable to track them down on GitHub, not sure whether that would be a problem. Many of these have also worked with the last author (@malsch) before on projects, not sure whether this would be an issue, but they usually match the thematic focus of the package (coding of occupations) very well.

- Matthias Schonlau
- Frederic Conrad
- Silvia Schwanhäuser
- Christian Kagerl
- Sebastian Hülle
- Arne Bethmann

Thank you for handling our submission!

Cheers,
Jan
