### 2023-05-31 #5366 @welch16
Hi Chris,

I would like to review the paper.

Thanks, Rene

From: Chris Vernon ***@***.***>
Date: Saturday, May 27, 2023 at 8:59 AM
Cc: Rene Welch Schwartz ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: occupationMeasurement: A Comprehensive Toolbox for Interactive Occupation Coding in Surveys (Issue #5366)

👋 @welch16<https://github.com/welch16> - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
