### 2023-05-31 #5505 @editorialbot
[REVIEW]: occupationMeasurement: A Comprehensive Toolbox for Interactive Occupation Coding in Surveys
**Submitting author:** @jansim (<a href="http://orcid.org/0000-0002-9406-7761">Jan Simson</a>)
**Repository:** https://github.com/occupationMeasurement/occupationMeasurement
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
