### 2023-06-26 occupationMeasurement/occupationMeasurement#6 @danielruss
Questions about JOSS paper
I am trying to recreate figure 1.   I followed the instructions in the one of the [vignettes --building your own](https://cran.r-project.org/web/packages/occupationMeasurement/vignettes/app-questionnaire.html#questionnaires) but I get an error:  **Error: multi-argument returns are not permitted.**

I was able to get your second example to work, and the "getting started" vignette seemed to work.

Is any project using occupationMeasurement?
~~How do users store the results?~~ Found the answer in the vignette.

The first two sentences of the paper are a little awkward. Consider changing it from a passive sentence, "There is ...", to a more active sentence "People earn a living..."  I would also drop the "easy-to-use" before "R-package" and "easily" before
"adapt".  The code is easy after going through the learning curve.

Since your paper does not give examples, I think referencing your vignettes would improve understanding where to get started.

I was unable to test the docker image (no access to docker).

Otherwise, I think the code addresses a significant problem, and with a few minor corrections should be ready for publication.
(https://github.com/openjournals/joss-reviews/issues/5505)
