### 2023-06-28 occupationMeasurement/occupationMeasurement#7 @welch16
Questions about tests for the JOSS review
Hi,

I downloaded the package and tried to reprocess the tests using `devtools::test()` or `shinytest2::test_app()`. In both cases, I get the same error stating that the "API logging is not working".

I am not really familiar with testing shiny apps, and I am not sure if I should have done something else before using those commands. I consider that it would be useful to have more detailed guidelines for users to contribute to the R package.

Otherwise, I think that the work is ready for publication,
https://github.com/openjournals/joss-reviews/issues/5505
