### 2023-07-03 7174637
Add missing `list()` wrap in questionnaire vignette (#6)
