+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-There is a multitude of ways to earn a living which is why the occupations people pursue are almost as diverse as people themselves. This makes quantitative analyses of free-text occupational responses from surveys hard to impossible, especially since people may refer to the same occupations with different terms. To address this problem, a variety of different classifications have been developed, such as the International Standard Classification of Occupations 2008 (ISCO) [@ilo2012] and the German Klassifikation der Berufe 2010 (KldB) [@bundesagenturfürarbeit2011], narrowing down the amount of occupation categories into more manageable numbers in the mid hundreds to low thousands and introducing a hierarchical ordering of categories. This leads to a different problem, however: Coding occupations into these standardized categories is usually expensive, time-intensive and plagued by issues of reliability.
+People earn a living a multitude of ways which is why the occupations they pursue are almost as diverse as people themselves. This makes quantitative analyses of free-text occupational responses from surveys hard to impossible, especially since people may refer to the same occupations with different terms. To address this problem, a variety of different classifications have been developed, such as the International Standard Classification of Occupations 2008 (ISCO) [@ilo2012] and the German Klassifikation der Berufe 2010 (KldB) [@bundesagenturfürarbeit2011], narrowing down the amount of occupation categories into more manageable numbers in the mid hundreds to low thousands and introducing a hierarchical ordering of categories. This leads to a different problem, however: Coding occupations into these standardized categories is usually expensive, time-intensive and plagued by issues of reliability.
 
 Here we present a new instrument that implements a faster, more convenient and interactive occupation coding workflow where respondents are included in the coding process. Based on the respondent's answer, a novel machine learning algorithm generates a list of suggested occupational categories from the [Auxiliary Classification of Occupations](https://github.com/occupationMeasurement/auxiliary-classification) [@schierholz2018], from which one is chosen by the respondent (see \autoref{fig:app_flow_screenshots}). Issues of ambiguity within occupational categories are addressed through clarifying follow-up questions. We provide a comprehensive toolbox including anonymized German training data and pre-trained models without raising privacy issues, something not possible yet with other algorithms due to the difficulties of anonymizing free-text data.
 
@@ -46,7 +46,7 @@ Our toolbox addresses these issues by implementing an occupation coding workflow
 
 # Functionalities
 
-We provide an open-source implementation of a machine learning algorithm for occupation coding with immediate feedback and verification, available as an easy-to-use R [@R] package on CRAN[^2]. To make it widely useful, our toolkit can be readily integrated in both self-administered web surveys as well as interviewer-administered (telephone) surveys using the included questionnaires. Programmers can easily adapt these questionnaires to fit a wide array of requirements. The toolbox includes custom survey software built on top of the shiny [@winston2023] framework to integrate machine learning predictions into surveys. On the off-chance that further flexibility is needed, we offer direct API access for completely custom data collection and integration into existing survey software. As we built the toolbox on top of the shiny [@winston2023] and plumber [@schloerke2022] frameworks, deployments on the Web are easy. We further provide [pre-built container images](https://github.com/orgs/occupationMeasurement/packages?repo_name=occupationMeasurement) for even easier deployment in production environments.
[…]
