### 2023-07-03 5371eef
Add further links to vignettes as per suggestions in #6
