+++ paper/paper.md
@@ -46,7 +46,7 @@ Our toolbox addresses these issues by implementing an occupation coding workflow
 
 # Functionalities
 
-We provide an open-source implementation of a machine learning algorithm for occupation coding with immediate feedback and verification, available as an R [@R] package on CRAN[^2]. To make it widely useful, our toolkit can be readily integrated in both self-administered web surveys as well as interviewer-administered (telephone) surveys using the included questionnaires. Programmers can adapt these questionnaires to fit a wide array of requirements. The toolbox includes custom survey software built on top of the shiny [@winston2023] framework to integrate machine learning predictions into surveys. On the off-chance that further flexibility is needed, we offer direct API access for completely custom data collection and integration into existing survey software. As we built the toolbox on top of the shiny [@winston2023] and plumber [@schloerke2022] frameworks, deployments on the Web are easy. We further provide [pre-built container images](https://github.com/orgs/occupationMeasurement/packages?repo_name=occupationMeasurement) for even easier deployment in production environments.
+We provide an open-source implementation of a machine learning algorithm for occupation coding with immediate feedback and verification, available as an R [@R] package on CRAN[^2]. An introductory ["Getting Started"](https://occupationmeasurement.github.io/occupationMeasurement/articles/occupationMeasurement.html) guide is available for anyone looking to use the package. To make it widely useful, our toolkit can be readily integrated in both self-administered web surveys as well as interviewer-administered (telephone) surveys using the included questionnaires. Programmers can [adapt these questionnaires](https://occupationmeasurement.github.io/occupationMeasurement/articles/app-questionnaire.html) to fit a wide array of requirements. The toolbox includes [custom survey software](https://occupationmeasurement.github.io/occupationMeasurement/articles/app.html) built on top of the shiny [@winston2023] framework to integrate machine learning predictions into surveys. On the off-chance that further flexibility is needed, we offer [direct API access](https://occupationmeasurement.github.io/occupationMeasurement/articles/api.html) for completely custom data collection and integration into existing survey software. As we built the toolbox on top of the shiny [@winston2023] and plumber [@schloerke2022] frameworks, [deployments on the Web](https://occupationmeasurement.github.io/occupationMeasurement/articles/app-deployment.html) are easy. We further provide [pre-built container images](https://github.com/orgs/occupationMeasurement/packages?repo_name=occupationMeasurement) for even easier deployment in production environments.
 
 The toolbox uses a specifically developed list of occupational task descriptions, the [Auxiliary Classification of Occupations](https://github.com/occupationMeasurement/auxiliary-classification), designed to be easier to understand and less ambiguous than existing lists of job titles [@schierholz2018]. Alongside this list, it provides matching follow-up questions to enable a fine-grained assignment into existing classification systems.
 
