### 2023-07-03 occupationMeasurement/occupationMeasurement#6 @jansim
Hi Daniel,
Thanks for your review and the great feedback! Sorry for the delay in my response, somehow I did not receive a notification of this issue.

I agree with the points you raised and adressed them as follows:

> I am trying to recreate figure 1. I followed the instructions in the one of the [vignettes --building your own](https://cran.r-project.org/web/packages/occupationMeasurement/vignettes/app-questionnaire.html#questionnaires) but I get an error: Error: multi-argument returns are not permitted.

Good catch, there was a missing `list()` wrap there. I've updated the vignette to fix this error. ✅

> Is any project using occupationMeasurement?

We were working with a survey institute, who successfully used the API implementation in one of their surveys and also informed some of the API design. Do you think we should mention this somewhere?

Regarding your feedback of the paper itself, I agree with your points and updated the paper to reflect your suggestions. I very much appreciate the helpful feedback!

All the best,
Jan
