### 2023-07-03 occupationMeasurement/occupationMeasurement#6 @danielruss
> We were working with a survey institute, who successfully used the API implementation in one of their surveys and also informed some of the API design. Do you think we should mention this somewhere?

Yes, I think you should mention it towards the end of the paper.  This lets people know that the code is good enough for use in a major study.
