### 2023-07-03 1b6d466
Improve debugging of api test case if api crashes
