### 2023-07-03 4278055
Improve detection of R packages installed via R CMD CHECK
This was erroneously detected when a packages was install from CRAN.
