+++ tests/testthat/test-api.R
@@ -19,7 +19,7 @@ package_info <- package_info[package_info$package == "occupationMeasurement", ]
 # will correspond to different things based on how the package was loaded.
 # For R CMD CHECK it will correspond to the built package which just needs
 # to be laoded, for devtools::test it will correspond to the package source
-installed_via_check <- package_info$path != "" && package_info$source != "load_all()"
+installed_via_check <- package_info$path != "" && package_info$source == "local"
 
 if (installed_via_check) {
   # Re-use the installation from R CMD CHECK and directly load
