### 2023-07-03 1e5ed01
Add new section on issues and contributing in README per suggestion in #7
