### 2023-07-03 occupationMeasurement/occupationMeasurement#7 @jansim
Hi Rene,

Thanks for your review and feedback! Good note on the tests, this was actually not related to any extra setup necessary, but rather a bug in the test code:

We need to handle the local package installation differently based on whether the test is triggered by R CMD CHECK (which installs a local version of the package already) or not (which means that we have to install a local version ourselves), this detection did not work correctly when the package was installed from CRAN, however. The issue should be resolved with [this commit](https://github.com/occupationMeasurement/occupationMeasurement/commit/42780555f813561044fb16de2a2e46b3381783ce), maybe you could try running the tests again?

I also like the idea of providing a bit more info / encouragement regarding contributions and added a brief note on this to the README.

Thanks for your suggestions, I very much appreciate the helpful feedback!

Best Regards,
Jan
