### 2023-07-03 bf2efb5
Remove development dependencies from `Suggessts` in `DESCRIPTION`
Fixes #5
