+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   app-questionnaire: app-questionnaire.html
   app: app.html
   occupationMeasurement: occupationMeasurement.html
-last_built: 2023-07-03T16:18Z
+last_built: 2023-07-03T16:34Z
 urls:
   reference: https://occupationmeasurement.github.io/occupationMeasurement/reference
   article: https://occupationmeasurement.github.io/occupationMeasurement/articles
