### 2023-07-10 occupationMeasurement/occupationMeasurement#7 @welch16
Hi Jan,

I tried it and now it works correctly! Great work!
