### 2023-07-14 occupationMeasurement/occupationMeasurement#6 @jansim
Will go ahead and close this issue, if you have additional comments / wish for some changes feel free to re-open this issue @danielruss. Thank you again for the great comments!
