### 2023-07-25 7a57000
Add most development dependencies back as `R CMD CHECK` wants them (CC #5)
