### 2023-07-25 occupationMeasurement/occupationMeasurement#8 @jansim
Release for openjournals/joss-reviews#5505
