+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   app-questionnaire: app-questionnaire.html
   app: app.html
   occupationMeasurement: occupationMeasurement.html
-last_built: 2023-07-25T14:07Z
+last_built: 2023-07-25T15:19Z
 urls:
   reference: https://occupationmeasurement.github.io/occupationMeasurement/reference
   article: https://occupationmeasurement.github.io/occupationMeasurement/articles
