+++ cran-comments.md
@@ -6,3 +6,10 @@
 
 This is a minor, incremental release with mainly QoL fixes.
 {styler} has been removed per request by its author from `Suggests` to reduce its number of reverse dependencies.
+
+## revdepcheck results
+
+We checked 0 reverse dependencies, comparing R CMD check results across CRAN and dev versions of this package.
+
+ * We saw 0 new problems
+ * We failed to check 0 packages
