### 2023-07-25 #5505 @jansim
Thanks @crvernon ! It seems that the CRAN team is on vacation, so I will only be able to make a new release after August 3rd, but will get back to this then 😊
