### 2023-08-04 69d78ee
Update KldB citation to keep capitalization
