+++ paper/paper.bib
@@ -178,7 +178,7 @@
 }
 
 @book{bundesagenturfürarbeit2011,
-	title = {Klassifikation der Berufe 2010},
+	title = {{Klassifikation der Berufe 2010}},
 	author = {{Bundesagentur für Arbeit}, },
 	year = {2011},
 	date = {2011},
