+++ tests/testthat/_snaps/suggestions.md
@@ -6,7 +6,7 @@
         "names": {
           "type": "character",
           "attributes": {},
-          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "has_followup_questions"]
+          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "notes", "has_followup_questions"]
         },
         "class": {
           "type": "character",
@@ -54,6 +54,11 @@
           "type": "character",
           "attributes": {},
           "value": ["Köche/Köchinnen", "Speisenzubereitung (Führungskraft)", "Köche/Köchinnen", "Fischverarbeitung", "Gastronomieservice (ohne Spezialisierung)"]
+        },
+        {
+          "type": "character",
+          "attributes": {},
+          "value": ["", "", "", "", ""]
         },
         {
           "type": "logical",
@@ -181,7 +186,7 @@
         "names": {
           "type": "character",
           "attributes": {},
-          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "has_followup_questions"]
+          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "notes", "has_followup_questions"]
         },
         "class": {
           "type": "character",
@@ -231,6 +236,11 @@
           "value": ["Metallerzeugung (ohne Spezialisierung)", "Fahrzeugtechnik", "Lebensmittelherstellung (ohne Spezialisierung)", "Hochbau, Tiefbau (ohne Spezialisierung)", "Kunststoff- und Kautschukherstellung"]
         },
         {
+          "type": "character",
+          "attributes": {},
+          "value": ["", "", "", "", ""]
+        },
+        {
           "type": "logical",
           "attributes": {},
           "value": [false, false, false, true, false]
@@ -246,7 +256,7 @@
         "names": {
           "type": "character",
           "attributes": {},
-          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "has_followup_questions"]
+          "value": ["auxco_id", "input_text", "score", "title", "task", "task_description", "kldb_title_short", "notes", "has_followup_questions"]
         },
         "class": {
           "type": "character",
@@ -294,6 +304,11 @@
           "type": "character",
           "attributes": {},
           "value": ["Studentische Hilfskraft", "Hochschullehre und -forschung", "Büro- und Sekretariatskräfte (ohne Spezialisierung)", "Büro- und Sekretariatskräfte", "Büro- und Sekretariatskräfte (ohne Spezialisierung)"]
+        },
+        {
+          "type": "character",
+          "attributes": {},
[…]
