### 2023-08-08 72a34f1
Update NEWS.md
