### 2023-08-08 a496dc4
Only run compute-heavy examples in interactive sessions
