### 2023-08-08 79ac332
Update URL to shiny tag glossary
Adressing a NOTE, which was probably a bug in CRAN, as another perfectly fine URL was also flagged
