### 2023-08-08 bd899a8
Remove reference to stringi, as it doesn't seem worth adding a revdep for them just for the one reference in the vignette.
