+++ R/helper_functions.R
@@ -4,7 +4,7 @@
 #' (e.g., Ä, Ü, "DIPL.-ING.") with their uppercase equivalents and removes
 #' punctuation, empty spaces and the word "Diplom".
 #'
-#' [charToRaw()] or [stringi::stri_escape_unicode()] helps to find UTF-8 characters.
+#' [charToRaw()] helps to find UTF-8 characters.
 #' @encoding UTF-8
 #' @param verbatim The character vector to process.
 #' @param lang The language the text is in.
