### 2023-08-11 #5505 @jansim
Hey @crvernon, sorry for the delay! I'm currently waiting to hear back from CRAN, regarding the new release and will hopefully be able to get it through soon. Will let you know ASAP once we hear back from them 👍️
