### 2023-08-16 occupationMeasurement/occupationMeasurement#8 @jansim
This release is currently held up on CRAN, by the following NOTE:

```
Flavor: r-devel-linux-x86_64-debian-gcc
Check: examples, Result: NOTE
  Examples with CPU time > 2.5 times elapsed time
       user system elapsed ratio
  app 4.281  0.326   1.206  3.82

Flavor: r-devel-linux-x86_64-debian-gcc
Check: tests, Result: NOTE
    Running 'testthat.R' [35s/4s]
  Running R code in 'testthat.R' had CPU time 7.9 times elapsed time
```

This is related to the package utilizing > 2 cores at the same time and CRAN recently changing their test server `env` variables so that other packages don't use a max of 2 cores by default anymore.

My guess here is that for our package this is mainly due to our use of `data.table` which uses multiple cores by default (although this is not the only package we use that uses OpenMP).

The current plan is too see how this is adressed by data.table (relevant issues: https://github.com/Rdatatable/data.table/issues/5658 [new] and https://github.com/Rdatatable/data.table/issues/3300 [old]) and then decide what to do. Overwriting the max number of threads in our package seems like odd behaviour / the wrong place to provide these settings, but we could add code to do that.
