### 2023-08-16 #5505 @jansim
Hey @crvernon ,
so we heard back from CRAN and after some investigation, the issue seems to be related to a recent change in their test environment settings in regards to parallel processing. Since the code that does the parallel processing actually lies in other packages (which we only happen to be using), I would suggest to make a release of the current version of the package right now and use this for the paper and we will then upload that release (or a slightly updated one) to CRAN later on once these other packages had some time to deal with the issue.

I wrote a more detailed description of the situation in the issue we use to track the new CRAN release: https://github.com/occupationMeasurement/occupationMeasurement/issues/8.

If you're OK with this, I can get the new Zenodo release done any time 😊
