### 2023-08-23 #5505 @jansim
Ahh so you're saying we can't rename the Zenodo archive?
