### 2023-08-23 #5505 @crvernon
That's right, your Zenodo archive title should match that of your paper.
