### 2023-08-23 #5505 @crvernon
So instead of:

> occupationMeasurement/occupationMeasurement: occupationMeasurement 0.3.0 (not on CRAN): JOSS release

You would need:

> occupationMeasurement: A Comprehensive Toolbox for Interactive Occupation Coding in Surveys
