### 2023-08-23 #5505 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
